dancecards: try to fix refresh bug
When a mobile device is in sleep mode for too long (more than a few minutes),
the socket.io connection will be re-established automatically, but the client
is no longer is the room "code" and also all changes to the stickers during
that time will be only reflected on refresh.

Hook into the "reconnect" event, re-join the room "code" and then go through
the list of all still empty sticker spots and ask the server to send an update
event for them.

Signed-off-by: Martin Kaistra <admin@djfun.de>