style: fix phone/email display in popup
The divs for phone number and email address in the popup have the css class
hidden, but are shown anyway, because there is no rule in the stylesheet for
that.

Add such a rule and add back the setting of the phone number in the popup. This
was previously commented out because phone number is not in the database for
all rallys.

Now, it can be made visible from the rally-specific stylesheet by setting
"display: unset" in #popup-message-email or #popup-message-phone.

Signed-off-by: Martin Kaistra <admin@djfun.de>